Fix script location
